+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Adds
 
 * Translation strings added for the layout- and layout-column-widgets.
+* New `box` schema field type
 * When switching locale from the doc editor, ask if the user wants to localize the current document in the target locale or want to start a blank document.
 * Introduced a new `longPolling: false` option for the `@apostrophecms/notification` module. This eliminates long-pending requests when logged in, but also slows down the delivery of notifications. The behavior can be tuned further via the `pollingInterval` option, which defaults to `5000` milliseconds.